requirements.txt
asgiref==3.5.0
boto3==1.20.46
botocore==1.23.46
Django==4.0.1
django-cors-headers==3.11.0
django-environ==0.8.1
django-filter==21.1
django-rest-framework==0.1.0
django-storages==1.12.3
djangorestframework==3.13.1
gunicorn==20.1.0
jmespath==0.10.0
Pillow==9.0.0
psycopg2==2.9.3
python-dateutil==2.8.2
pytz==2021.3
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8
whitenoise==5.3.0
